Make process not reserving stack for handle_trap
Previously, when handle_trap was called by kernel_entry, it consumed
 stack space.
To account for this, create_process had to subtract 16 from sp so it
 could later be restored.
However, this coupling meant that any future changes to handle_trap
could require adjusting that subtraction value.

This change modifies switch_context so that the value stored in
current->sp is the one before handle_trap is invoked,
rather than the current sp.
As a result, handle_trap can now simply return without restoring sp.